% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/GetGeo.R
\name{getGeo}
\alias{getGeo}
\title{Geocodes a geographical location and returns it latitude and longitude}
\usage{
getGeo(addr)
}
\arguments{
\item{addr}{character vector containing the location to geocode}
}
\description{
Geocodes a geographical location and returns it latitude and longitude
}
\examples{
getGeo("Weststr. 88, 33615 Bielefeld")
getGeo("Berlin, Alexanderplatz")
}
\keyword{API}
\keyword{Google}
\keyword{geocode}
\keyword{geodata}
\keyword{location}

